rhcos/v0_2_exp: rename to openshift/v4_8_exp
We want to support both OCP and OKD.
