+++ docs/config-openshift-v4_8-exp.md
@@ -1,18 +1,18 @@
 ---
 layout: default
-title: RHEL CoreOS v0.2.0-experimental
+title: OpenShift v4.8.0-experimental
 parent: Configuration specifications
 nav_order: 100
 ---
 
-# RHEL CoreOS Specification v0.2.0-experimental
+# OpenShift Specification v4.8.0-experimental
 
 **Note: This configuration is experimental and has not been stabilized. It is subject to change without warning or announcement.**
 
-The RHEL CoreOS configuration is a YAML document conforming to the following specification, with **_italicized_** entries being optional:
+The OpenShift configuration is a YAML document conforming to the following specification, with **_italicized_** entries being optional:
 
 * **variant** (string): used to differentiate configs for different operating systems. Must be `rhcos` for this specification.
-* **version** (string): the semantic version of the spec for this document. This document is for version `0.2.0-experimental` and generates Ignition configs with version `3.3.0-experimental`.
+* **version** (string): the semantic version of the spec for this document. This document is for version `4.8.0-experimental` and generates Ignition configs with version `3.3.0-experimental`.
 * **ignition** (object): metadata about the configuration itself.
   * **_config_** (objects): options related to the configuration.
     * **_merge_** (list of objects): a list of the configs to be merged to the current config.
